moved shoot sort and search methods to utils
